move FormToggle as shared component
