+++ frontend/src/components/FormToggle/View.tsx
