+++ frontend/src/pages/ProjectDetails/Payments/PaymentForm/WorkItemSidePanel/OtherWorkForm/Description.tsx
@@ -1,7 +1,7 @@
 import classNames from "classnames";
 import { useForm, useWatch } from "react-hook-form";
 import Input from "src/components/FormInput";
-import FormToggle from "./FormToggle";
+import FormToggle from "src/components/FormToggle";
 import MarkdownPreview from "src/components/MarkdownPreview";
 import { useIntl } from "src/hooks/useIntl";
